[NC] Add contraint change factors, movement dampening and variable axis locking (#66)
Change factors
Multiply the movement/rotation/scale change of the parent by a factor and apply that movement to the child. Defaults to 1 to keep current behaviour.

e.g. with a movement factor of 2 on the position, the following would occur:
parent: (0, 0, 0) -> (3, 6, 0)
child: (0, 0, 0) -> (6, 12, 0)

Movement dampening
Dampens the childs movement towards the parent, making it follow the parent instead of a 1:1 position lock.
0 (or smaller) disables this and keeps current behaviour (and is the default).
The higher the number, the faster the child will follow the parent.

Axis locking
Only constrain the chosen axis to the parent. Letting the child move freely on the non-constrained axis.
The X/Y/Z labels are replaced by toggles styled as buttons, and are all turned on by default to keep current behaviour.

Added an extra row for every transform where I now put all the new functions