QoL update Colliders Editor | Force refresh colliders (#141)
I was having issues with colliders not being added to the colliders list properly when a new character was added to a scene, the colliders were only getting added to the list properly when you opened an existing scene with characters, only those that were already there had the full list. This manually fixes the issue to anybody that requires a refresh.